chore: move delegations lookup endpoint into pox router
