+++ src/api/routes/pox.ts
@@ -3,13 +3,21 @@ import { asyncHandler } from '../async-handler';
 import { getPagingQueryLimit, parsePagingQueryInput, ResourceType } from '../pagination';
 import { PgStore } from '../../datastore/pg-store';
 import { parsePoxSyntheticEvent } from '../controllers/db-controller';
-import { validatePrincipal, validateRequestHexInput } from '../query-helpers';
+import {
+  getBlockHeightQueryParam,
+  getBlockParams,
+  validatePrincipal,
+  validateRequestHexInput,
+} from '../query-helpers';
+import { getETagCacheHandler, setETagCacheHeaders } from '../controllers/cache-controller';
+import { PoolDelegationsResponse } from '@stacks/stacks-blockchain-api-types';
 
 export function createPoxEventsRouter(
   db: PgStore,
   poxVersion: 'pox2' | 'pox3' | 'pox4'
 ): express.Router {
   const router = express.Router();
+  const cacheHandler = getETagCacheHandler(db);
 
   const poxTable = (
     {
@@ -21,6 +29,7 @@ export function createPoxEventsRouter(
 
   router.get(
     '/events',
+    cacheHandler,
     asyncHandler(async (req, res) => {
       const limit = getPagingQueryLimit(ResourceType.Pox2Event, req.query.limit);
       const offset = parsePagingQueryInput(req.query.offset ?? 0);
@@ -36,12 +45,14 @@ export function createPoxEventsRouter(
         offset,
         results: parsedResult,
       };
+      setETagCacheHeaders(res);
       res.json(response);
     })
   );
 
   router.get(
     '/tx/:tx_id',
+    cacheHandler,
     asyncHandler(async (req, res) => {
       const { tx_id } = req.params;
       validateRequestHexInput(tx_id);
@@ -57,12 +68,14 @@ export function createPoxEventsRouter(
       const response = {
         results: parsedResult,
       };
+      setETagCacheHeaders(res);
       res.json(response);
     })
   );
 
   router.get(
     '/stacker/:principal',
+    cacheHandler,
     asyncHandler(async (req, res) => {
       const { principal } = req.params;
       validatePrincipal(principal);
@@ -78,6 +91,66 @@ export function createPoxEventsRouter(
       const response = {
         results: parsedResult,
       };
+      setETagCacheHeaders(res);
+      res.json(response);
+    })
+  );
+
+  router.get(
+    '/:pool_principal/delegations',
+    cacheHandler,
+    asyncHandler(async (req, res, next) => {
+      // get recent asset event associated with address
+      const poolPrincipal = req.params['pool_principal'];
+      validatePrincipal(poolPrincipal);
+
+      const limit = getPagingQueryLimit(ResourceType.Stacker, req.query.limit);
+      const offset = parsePagingQueryInput(req.query.offset ?? 0);
+      const afterBlock = getBlockHeightQueryParam('after_block', false, req, res, next) || 0;
+
+      const response = await db.sqlTransaction(async sql => {
+        const blockParams = getBlockParams(req, res, next);
+        let blockHeight: number;
+        if (blockParams.blockHeight !== undefined) {
+          blockHeight = blockParams.blockHeight;
+        } else {
+          blockHeight = await db.getMaxBlockHeight(sql, {
+            includeUnanchored: blockParams.includeUnanchored ?? false,
+          });
+        }
+
+        const dbBlock = await db.getBlockByHeightInternal(sql, blockHeight);
+        if (!dbBlock.found) {
+          const error = `no block at height: ${blockHeight}`;
+          res.status(404).json({ error: error });
+          throw new Error(error);
+        }
+        const burnBlockHeight = dbBlock.result.burn_block_height;
+
+        const stackersQuery = await db.getPoxPoolDelegations({
+          delegator: poolPrincipal,
+          blockHeight,
+          burnBlockHeight,
+          afterBlockHeight: afterBlock,
+          limit,
+          offset,
+          poxTable,
+        });
+        if (!stackersQuery.found) {
+          const error = `no stackers found`;
+          res.status(404).json({ error: error });
+          throw new Error(error);
+        }
+
+        const response: PoolDelegationsResponse = {
+          limit,
+          offset,
+          total: stackersQuery.result.total,
+          results: stackersQuery.result.stackers,
+        };
+        return response;
+      });
+      setETagCacheHeaders(res);
       res.json(response);
     })
   );
